third_party/WebKit/LayoutTests/virtual/stable/animations-unprefixed/frames-timing-function.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#target {
  transition-timing-function: frames(4);
  animation-timing-function: frames(4);
}
</style>

<div id="target"></div>

<script>
test(function() {
  assert_equals(getComputedStyle(target).transitionTimingFunction, 'ease');
  assert_equals(getComputedStyle(target).animationTimingFunction, 'ease');
}, 'The frames() timing function is still experimental and should not be exposed to stable.');
</script>
